requirements.txt
annotated-types==0.6.0
asttokens==2.4.1
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
Eel==0.16.0
exceptiongroup==1.2.1
executing==2.0.1
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
importlib-metadata==7.1.0
jedi==0.19.1
Markdown==3.6
matplotlib-inline==0.1.7
numpy==1.26.4
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic-core==2.18.2
pygments==2.18.0
pyparsing==3.1.2
requests==2.31.0
rsa==4.9
six==1.16.0
stack-data==0.6.3
tqdm==4.66.4
traitlets==5.14.3
typing-extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.6
wcwidth==0.2.13
whichcraft==0.6.1
wxPython==4.2.1
zipp==3.18.1
zope.event==5.0
zope.interface==6.3
